requirements.txt
Flask==3.1.0
numpy>=1.26.0
google-genai>=1.30.0
